Move the func_template under scripts/
